+++ src/js/actions/clearModalAction.js
@@ -1,0 +1,11 @@
+
+export default () => {
+    $('#modalCart')
+        .on('hidden.bs.modal', function (e) {
+            console.log("hidden.bs.modal");
+
+            $('#modal-content-section')
+                .html("");
+
+        });
+}
