+++ src/js/actions/initCardAction.js
@@ -1,0 +1,8 @@
+export default () => {
+    if(
+        !localStorage.getItem("cart") ||
+        !Array.isArray(JSON.parse(localStorage.getItem("cart")))
+    ){
+        localStorage.setItem("cart",JSON.stringify([]));
+    }
+}
