+++ src/js/actions/showCardModalAction.js
@@ -1,0 +1,16 @@
+import createCardList from "../components/CardItemsList";
+
+export default () => {
+    document.getElementById("cart_nav")
+        .addEventListener("click",()=>{
+
+            if(
+                localStorage.getItem("cart") &&
+                Array.isArray(JSON.parse(localStorage.getItem("cart")))
+            ){
+                const cardItems = JSON.parse(localStorage.getItem("cart"));
+                createCardList(cardItems);
+            }
+
+        });
+}
