+++ src/js/index.js
@@ -5,9 +5,11 @@ import header from './components/Header';
 import main from './components/Main';
 import modal from './components/ModalWindow';
 
-import createCardList from './components/CardItemsList'
 
 import addToCartAction from './actions/addToCartAction';
+import showCardModalAction from './actions/showCardModalAction';
+import clearModalAction from './actions/clearModalAction';
+import initCardAction from './actions/initCardAction';
 
 const _header = header();
 const _main = main();
@@ -27,37 +29,9 @@ render("app", _header);
 render("app", _main);
 render("app", modal);
 
-if(
-    !localStorage.getItem("cart") ||
-    !Array.isArray(JSON.parse(localStorage.getItem("cart")))
-){
-    localStorage.setItem("cart",JSON.stringify([]));
-}
-
-[...document.getElementsByClassName("nav-item")]
-    .map(el => el.addEventListener("click", () => console.log("nav-item")));
-
+initCardAction();
 addToCartAction(addToCartEvent);
+showCardModalAction();
+clearModalAction();
 
-document.getElementById("cart_nav")
-    .addEventListener("click",()=>{
 
-        if(
-            localStorage.getItem("cart") &&
-            Array.isArray(JSON.parse(localStorage.getItem("cart")))
-        ){
-           const cardItems = JSON.parse(localStorage.getItem("cart"));
-           // document.getElementById("modal-content-section")
-            console.log(createCardList(cardItems));
-        }
-
-    });
-
-$('#modalCart')
-    .on('hidden.bs.modal', function (e) {
-    console.log("hidden.bs.modal");
-
-    $('#modal-content-section')
-        .html("");
-
-});
